+++ features/fixtures/delayed_job/app/app/jobs/test_report_context_job.rb
@@ -1,0 +1,7 @@
+class TestReportContextJob < ApplicationJob
+  queue_as :default
+
+  def perform(*)
+    raise "oh dear"
+  end
+end